+++ nano/store/ledger/version.hpp
@@ -9,8 +9,8 @@ class version
 public:
 	explicit version (store::backend &);
 
-	void put (store::write_transaction const & tx, int version);
-	int get (store::transaction const & tx) const;
+	void put (store::write_transaction const & tx, uint64_t version);
+	uint64_t get (store::transaction const & tx) const;
 
 private:
 	store::backend & backend;
